refac: move ExpenseBriefCards class to other file
